use uint32_t instead of uint64_t & add pragma unroll back

+++ include/flashinfer/layout.cuh
@@ -28,21 +28,23 @@ __host__ __device__ __forceinline__ size_t get_elem_offset_impl(size_t elem_idx,
 
 template <QKVLayout layout>
 struct tensor_info_t {
-  size_t qo_len;
-  size_t kv_len;
-  size_t num_heads;
-  size_t head_dim;
-  __host__ __device__ __forceinline__ tensor_info_t(size_t qo_len, size_t kv_len, size_t num_heads,
-                                                    size_t head_dim)
+  uint32_t qo_len;
+  uint32_t kv_len;
+  uint32_t num_heads;
+  uint32_t head_dim;
+  __host__ __device__ __forceinline__ tensor_info_t(uint32_t qo_len, uint32_t kv_len,
+                                                    uint32_t num_heads, uint32_t head_dim)
       : qo_len(qo_len), kv_len(kv_len), num_heads(num_heads), head_dim(head_dim) {}
 
-  __host__ __device__ __forceinline__ size_t get_qo_elem_offset(size_t query_idx, size_t head_idx,
-                                                                size_t feat_idx) const {
+  __host__ __device__ __forceinline__ uint32_t get_qo_elem_offset(uint32_t query_idx,
+                                                                  uint32_t head_idx,
+                                                                  uint32_t feat_idx) const {
     return get_elem_offset_impl<layout>(query_idx, head_idx, feat_idx, qo_len, num_heads, head_dim);
   }
 
-  __host__ __device__ __forceinline__ size_t get_kv_elem_offset(size_t kv_idx, size_t head_idx,
-                                                                size_t feat_idx) const {
+  __host__ __device__ __forceinline__ uint32_t get_kv_elem_offset(uint32_t kv_idx,
+                                                                  uint32_t head_idx,
+                                                                  uint32_t feat_idx) const {
     return get_elem_offset_impl<layout>(kv_idx, head_idx, feat_idx, kv_len, num_heads, head_dim);
   }
 };
