+++ include/flashinfer/permuted_smem.cuh
@@ -16,7 +16,7 @@ namespace flashinfer {
 using bank_t = uint4;
 
 template <typename T>
-constexpr __host__ __device__ __forceinline__ size_t bank_capacity() {
+constexpr __host__ __device__ __forceinline__ uint32_t bank_capacity() {
   return sizeof(bank_t) / sizeof(T);
 }
 
@@ -31,84 +31,85 @@ namespace permuted_smem_impl {
  * \note The permuted shared memory maps 8x4 block in logical space to 4x8 block in physical space.
  * \see GTC 2020: Developing CUDA Kernels to Push Tensor Cores to the Absolute Limit on NVIDIA A100.
  */
-template <size_t stride, typename T>
-__host__ __device__ __forceinline__ bank_t *get_smem_ptr(T *smem_base, size_t i, size_t j) {
+template <uint32_t stride, typename T>
+__host__ __device__ __forceinline__ bank_t *get_smem_ptr(T *smem_base, uint32_t i, uint32_t j) {
   return ((bank_t *)smem_base) + (i / 2) * stride * 2 + (j / 4) * 8 + (i % 2) * 4 +
          ((j % 4) ^ ((i / 2) % 4));
 }
 
-template <size_t stride, typename T>
-__device__ __forceinline__ void ldmatrix_m8n8x4(uint32_t *R, T *smem_base, size_t i, size_t j) {
+template <uint32_t stride, typename T>
+__device__ __forceinline__ void ldmatrix_m8n8x4(uint32_t *R, T *smem_base, uint32_t i, uint32_t j) {
   bank_t *smem_ptr = get_smem_ptr<stride>(smem_base, i, j);
   mma::ldmatrix_m8n8x4(R, smem_ptr);
 }
 
-template <size_t stride, typename T>
-__device__ __forceinline__ void ldmatrix_m8n8x4_trans(uint32_t *R, T *smem_base, size_t i,
-                                                      size_t j) {
+template <uint32_t stride, typename T>
+__device__ __forceinline__ void ldmatrix_m8n8x4_trans(uint32_t *R, T *smem_base, uint32_t i,
+                                                      uint32_t j) {
   bank_t *smem_ptr = get_smem_ptr<stride>(smem_base, i, j);
   mma::ldmatrix_m8n8x4_trans(R, smem_ptr);
 }
 
-template <size_t stride, typename T>
-__device__ __forceinline__ void stmatrix_m8n8x4(uint32_t *R, T *smem_base, size_t i, size_t j) {
+template <uint32_t stride, typename T>
+__device__ __forceinline__ void stmatrix_m8n8x4(uint32_t *R, T *smem_base, uint32_t i, uint32_t j) {
   bank_t *smem_ptr = get_smem_ptr<stride>(smem_base, i, j);
   mma::stmatrix_m8n8x4(R, smem_ptr);
 }
 
-template <size_t stride, typename T>
-__device__ __forceinline__ void load_bank(T *smem_base, size_t i, size_t j, const T *gptr) {
+template <uint32_t stride, typename T>
+__device__ __forceinline__ void load_bank(T *smem_base, uint32_t i, uint32_t j, const T *gptr) {
   *get_smem_ptr<stride>(smem_base, i, j) = *reinterpret_cast<const bank_t *>(gptr);
 }
 
-template <size_t stride, typename T>
-__device__ __forceinline__ void store_bank(T *smem_base, size_t i, size_t j, T *gptr) {
+template <uint32_t stride, typename T>
+__device__ __forceinline__ void store_bank(T *smem_base, uint32_t i, uint32_t j, T *gptr) {
   *reinterpret_cast<bank_t *>(gptr) = *get_smem_ptr<stride>(smem_base, i, j);
 }
 
-template <size_t stride, typename T>
-__device__ __forceinline__ void load_bank_async(T *smem_base, size_t i, size_t j, const T *gptr,
+template <uint32_t stride, typename T>
+__device__ __forceinline__ void load_bank_async(T *smem_base, uint32_t i, uint32_t j, const T *gptr,
                                                 bool predicate) {
   bank_t *smem_ptr = get_smem_ptr<stride>(smem_base, i, j);
   cp_async::pred_load_128<true>(smem_ptr, reinterpret_cast<const bank_t *>(gptr), predicate);
 }
 
-template <size_t stride, typename T>
-__device__ __forceinline__ void load_bank_async(T *smem_base, size_t i, size_t j, const T *gptr) {
+template <uint32_t stride, typename T>
+__device__ __forceinline__ void load_bank_async(T *smem_base, uint32_t i, uint32_t j,
+                                                const T *gptr) {
   bank_t *smem_ptr = get_smem_ptr<stride>(smem_base, i, j);
   cp_async::load_128<true>(smem_ptr, reinterpret_cast<const bank_t *>(gptr));
 }
 
 }  // namespace permuted_smem_impl
 
-template <size_t stride, typename T>
+template <uint32_t stride, typename T>
 struct permuted_smem_t {
   T __align__(16) * base;
   __device__ __forceinline__ permuted_smem_t() : base(nullptr) {}
   __device__ __forceinline__ permuted_smem_t(T *base) : base(base) {}
-  __device__ __forceinline__ bank_t *get_ptr(size_t i, size_t j) {
+  __device__ __forceinline__ bank_t *get_ptr(uint32_t i, uint32_t j) {
     return permuted_smem_impl::get_smem_ptr<stride>(base, i, j);
   }
-  __device__ __forceinline__ void ldmatrix_m8n8x4(uint32_t *R, size_t i, size_t j) {
+  __device__ __forceinline__ void ldmatrix_m8n8x4(uint32_t *R, uint32_t i, uint32_t j) {
     permuted_smem_impl::ldmatrix_m8n8x4<stride>(R, base, i, j);
   }
-  __device__ __forceinline__ void ldmatrix_m8n8x4_trans(uint32_t *R, size_t i, size_t j) {
+  __device__ __forceinline__ void ldmatrix_m8n8x4_trans(uint32_t *R, uint32_t i, uint32_t j) {
     permuted_smem_impl::ldmatrix_m8n8x4_trans<stride>(R, base, i, j);
   }
-  __device__ __forceinline__ void stmatrix_m8n8x4(uint32_t *R, size_t i, size_t j) {
+  __device__ __forceinline__ void stmatrix_m8n8x4(uint32_t *R, uint32_t i, uint32_t j) {
     permuted_smem_impl::stmatrix_m8n8x4<stride>(R, base, i, j);
   }
-  __device__ __forceinline__ void load_bank(size_t i, size_t j, const T *gptr) {
+  __device__ __forceinline__ void load_bank(uint32_t i, uint32_t j, const T *gptr) {
     permuted_smem_impl::load_bank<stride>(base, i, j, gptr);
   }
-  __device__ __forceinline__ void load_bank_async(size_t i, size_t j, const T *gptr,
+  __device__ __forceinline__ void load_bank_async(uint32_t i, uint32_t j, const T *gptr,
                                                   bool predicate) {
     permuted_smem_impl::load_bank_async<stride>(base, i, j, gptr, predicate);
   }
-  __device__ __forceinline__ void load_bank_async(size_t i, size_t j, const T *gptr) {
+  __device__ __forceinline__ void load_bank_async(uint32_t i, uint32_t j, const T *gptr) {
     permuted_smem_impl::load_bank_async<stride>(base, i, j, gptr);
   }
-  __device__ __forceinline__ void store_bank(size_t i, size_t j, T *gptr) {
+  __device__ __forceinline__ void store_bank(uint32_t i, uint32_t j, T *gptr) {
     permuted_smem_impl::store_bank<stride>(base, i, j, gptr);
   }
 };
